s02e05/requirements.txt
openai>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
html2text>=2020.1.16
loguru>=0.7.0
python-dotenv>=1.0.0
tiktoken>=0.5.0
Pillow>=10.0.0  # dla obsługi obrazów
faster-whisper>=0.9.0  # dla transkrypcji audio
torch>=2.0.0  # wymagane przez faster-whisper
pydub>=0.25.1  # dla obsługi plików audio 